+++ src/flow/platform/style.native.js
@@ -1,0 +1,6 @@
+// @flow
+
+type ReactNativeStyleSheetId = number
+type ReactNativeStyleObject = { [string]: any }
+
+export type ReactNativePredefinedStyle = ReactNativeStyleSheetId | ReactNativeStyleObject