+++ src/flow/platform/style.web.js
@@ -1,0 +1,5 @@
+// @flow
+
+type CSSClassName = string
+
+export type WebPredefinedStyle = CSSClassName